requirements.txt
flask
google-genai
python-dotenv
python-owasp-zap-v2.4
